addresses comments. Moves edi helper to Translator
